+++ lib/base/bonfire_game.dart
@@ -34,7 +34,7 @@ class BonfireGame extends CustomBaseGame with KeyboardEvents {
   /// Represents the character controlled by the user in the game. Instances of this class has actions and movements ready to be used and configured.
   final Player? player;
 
-  /// The way you cand raw things like life bars, stamina and settings. In another words, anything that you may add to the interface to the game.
+  /// The way you can draw things like life bars, stamina and settings. In another words, anything that you may add to the interface to the game.
   final GameInterface? interface;
 
   /// Represents a map (or world) where the game occurs.
@@ -67,10 +67,10 @@ class BonfireGame extends CustomBaseGame with KeyboardEvents {
   /// Used to show in the interface the FPS.
   final bool showFPS;
 
-  Iterable<Lighting> _visibleLights = [];
-  Iterable<GameComponent> _visibleComponents = [];
-  Iterable<ObjectCollision> _visibleCollisions = [];
-  Iterable<ObjectCollision> _collisions = [];
+  Iterable<Lighting> _visibleLights = List.empty();
+  Iterable<GameComponent> _visibleComponents = List.empty();
+  Iterable<ObjectCollision> _visibleCollisions = List.empty();
+  Iterable<ObjectCollision> _collisions = List.empty();
   IntervalTick? _interval;
   ColorFilterComponent _colorFilterComponent = ColorFilterComponent(
     GameColorFilter(),
@@ -241,22 +241,27 @@ class BonfireGame extends CustomBaseGame with KeyboardEvents {
   void _updateTempList() {
     _visibleComponents = components.where((element) {
       return (element is GameComponent) && (element).isVisibleInCamera();
-    }).cast();
+    }).cast()
+      ..toList(growable: false);
 
     _collisions = components.where((element) {
       return (element is ObjectCollision) && (element).containCollision();
-    }).cast();
+    }).cast()
+      ..toList(growable: false);
 
     _visibleCollisions = _visibleComponents.where((element) {
       return (element is ObjectCollision) && (element).containCollision();
-    }).cast();
+    }).cast()
+      ..toList(growable: false);
 
-    _visibleCollisions = _visibleCollisions.toList()
-      ..addAll(map.getCollisionsRendered());
+    _visibleCollisions = (_visibleCollisions.toList()
+          ..addAll(map.getCollisionsRendered()))
+        .toList(growable: false);
 
     _visibleLights = components.where((element) {
       return element is Lighting && element.isVisible(camera);
-    }).cast();
+    }).cast()
+      ..toList(growable: false);
 
     gameController?.notifyListeners();
   }