+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.4.11
+- improvements in `MapWorld`
+- enables remove tiles of map.
+- fix position translation on diagonal movement of FlyingAttackObject [#155](https://github.com/RafaelBarbosatec/bonfire/pull/155)
+
 # 1.4.10
 - improvements performance